# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Fixture Platform API (DEV)
  version: 0.1.0
  description: Phase 1A — User types, Org membership, RBAC modules, Super-admin console.
paths:
...
  /api/permissions/orgs/{org_uuid}/users/{user_uuid}/grants/:
    get:
      operationId: permissions_orgs_users_grants_list
      description: List all grant rows + computed default-on modules for a user in
        an org.
      parameters:
      - in: path
        name: org_uuid
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: user_uuid
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - permissions
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GrantRow'
          description: ''
    put:
      operationId: permissions_orgs_users_grants_update
      description: 'Replace grants atomically. Accepts EITHER `{grants: [{module,
        state}], reason}` (existing shape) OR `{cells: {module_code: state}, reason,
        event_id?}` (SPA matrix shape). If both keys are present, `cells` wins.'
      parameters:
      - in: path
        name: org_uuid
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: user_uuid
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - permissions
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkGrants'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/BulkGrants'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/BulkGrants'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GrantRow'
          description: ''
...
  /api/permissions/orgs/{slug}/users/{user_uuid}/grants/:
    get:
      operationId: permissions_orgs_users_grants_list_2
      description: Slug-routed alias for GET /grants/ — same response shape.
      parameters:
      - in: path
        name: slug
        schema:
          type: string
        required: true
      - in: path
        name: user_uuid
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - permissions
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GrantRow'
          description: ''
    put:
      operationId: permissions_orgs_users_grants_update_2
      description: 'Slug-routed alias for PUT /grants/. Accepts EITHER the `{grants:
        [{module, state}], reason}` shape or the SPA matrix `{cells: {module: state},
        reason, event_id?}` shape.'
      parameters:
      - in: path
        name: slug
        schema:
          type: string
        required: true
      - in: path
        name: user_uuid
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - permissions
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkGrants'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/BulkGrants'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/BulkGrants'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GrantRow'
          description: ''
components:
  schemas:
...
    BulkGrants:
      type: object
      description: PUT body for /grants/ — replace all grants atomically.
      properties:
        grants:
          type: array
          items:
            $ref: '#/components/schemas/GrantInput'
        reason:
          type: string
          maxLength: 2000
          minLength: 20
      required:
      - grants
      - reason
...
    GrantInput:
      type: object
      description: Single (module, state) pair used inside the bulk grants PUT body.
      properties:
        module:
          type: string
          maxLength: 64
        state:
          $ref: '#/components/schemas/StateEnum'
      required:
      - module
      - state
    GrantRow:
      type: object
      description: A single MembershipModuleGrant row, with module code denormalized.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        module_code:
          type: string
          readOnly: true
        state:
          allOf:
          - $ref: '#/components/schemas/StateEnum'
          readOnly: true
        reason:
          type: string
          readOnly: true
        granted_by:
          type: string
          format: uuid
          readOnly: true
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - granted_by
      - id
      - module_code
      - reason
      - state
      - updated_at
...
    StateEnum:
      enum:
      - default
      - grant
      - deny
      type: string
      description: |-
        * `default` - Default (no override)
        * `grant` - Grant (force on)
        * `deny` - Deny (force off)
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid